Stop showing all of the performance plots
There are now over 1000 PNG plots being generated during the nightly
performance jobs. This makes the build page WAY too big and take really
long to load. Even after it does, finding the plot you're interested in
is almost impossible given that there are no text titles to search for
(the titles are embedded in the images).

I think we should just disable the image gallery plugin. We can keep the
plots as archived file artifacts so that they can be linked to from the
build-over-build plots.