Pad ctest regex argument with a space for shell quoting.
This is a temporay workaround to deal with some odd shell quoting
behavior in the build farm scripts stack. The test args substring
`(xfail|ikos)` is quoted since it contains shell control characters but
the quotes are being stripped from the string at some point when being
passed through the ros_buildfarm scripts (which could be happening in a
layer outside of our direct authorship) despite working as expected in a
neigboring substring `"not xfail and not ikos"`.

Scott astutely observed that the latter substring contains spaces while
the problematic regex does not. Padding the regex with an optional
trailing space, i.e. `(xfail|ikos) ?` is sufficient to prevent the
stripping of quotes from the substring and allow the job to run without
shell errors.